backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# ML / AI
scikit-learn>=1.4.0
xgboost>=2.0.0
pandas>=2.0.0
numpy>=1.26.0

# HuggingFace (Phase 3) — optional, falls back to NB if absent
transformers>=4.39.0
torch>=2.2.0

# Computer Vision (Phase 3 trailer pipeline) — optional
opencv-python-headless>=4.9.0.80
pytesseract>=0.3.10

# Data persistence
pymongo>=4.7.0
python-dotenv>=1.0.0

# Phase 6: Real-Time Data APIs
pytrends>=4.9.2                     # Google Trends (unofficial)
tweepy>=4.14.0                       # Twitter/X API v2
google-api-python-client>=2.127.0   # YouTube Data API v3

# Phase 6: Scheduler
APScheduler>=3.10.4                  # Hourly cron jobs

# Phase 7: Production / Testing
gunicorn>=21.2.0
pytest>=8.0.0
httpx>=0.27.0

# Audio (optional — trailer audio analysis)
librosa>=0.10.1
